refactor: rename style to core so pylint doesn't complain anymore
