+++ src/nitpick/project.py
@@ -216,7 +216,7 @@ class Project:
         config = self.read_configuration()
 
         # pylint: disable=import-outside-toplevel
-        from nitpick.style.style import Style
+        from nitpick.style.core import Style
 
         style = Style(self, offline, config.cache)
         style_errors = list(style.find_initial_styles(peekable(always_iterable(config.styles))))
@@ -242,7 +242,7 @@ class Project:
 
     def create_configuration(self) -> None:
         """Create a configuration file."""
-        from nitpick.style.style import Style  # pylint: disable=import-outside-toplevel
+        from nitpick.style.core import Style  # pylint: disable=import-outside-toplevel
 
         doc = document()
         doc.add(comment("This file was generated by the `nitpick init` command"))
