+++ src/nitpick/style/__init__.py
@@ -1,1 +1,5 @@
 """Styles parsing and merging."""
+from .cache import parse_cache_option
+from .core import Style
+
+__all__ = ("Style", "parse_cache_option")
